+++ tests/NCronJob.Tests/RuntimeJobRegistryTests.cs
@@ -1,0 +1,307 @@
+using System.Threading.Channels;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Time.Testing;
+using Shouldly;
+
+namespace NCronJob.Tests;
+
+public class RuntimeJobRegistryTests : JobIntegrationBase
+{
+    [Fact]
+    public async Task DynamicallyAddedJobIsExecuted()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob();
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.AddJob(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *"));
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(1);
+        jobFinished.ShouldBeTrue();
+    }
+
+    [Fact]
+    public async Task MultipleDynamicallyAddedJobsAreExecuted()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob();
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.AddJob(s => s
+            .AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *")
+            .AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *"));
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(2);
+        jobFinished.ShouldBeTrue();
+    }
+
+    [Fact]
+    public async Task CanRemoveJobByName()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(
+            s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *", jobName: "Job"));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.RemoveJob("Job");
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(1, TimeSpan.FromMilliseconds(200));
+        jobFinished.ShouldBeFalse();
+    }
+
+    [Fact]
+    public async Task CanRemoveByJobType()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *")));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.RemoveJob<SimpleJob>();
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(1, TimeSpan.FromMilliseconds(200));
+        jobFinished.ShouldBeFalse();
+    }
+
+    [Fact]
+    public async Task CanUpdateScheduleOfAJob()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("0 0 * * *").WithName("JobName")));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.UpdateSchedule("JobName", "* * * * *");
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(1);
+        jobFinished.ShouldBeTrue();
+    }
+
+    [Fact]
+    public async Task ShouldThrowAnExceptionWhenJobIsNotFoundAndTryingToUpdateSchedule()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob();
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        Should.Throw<InvalidOperationException>(() => registry.UpdateSchedule("JobName", "* * * * *"));
+    }
+
+    [Fact]
+    public async Task ShouldRetrieveScheduleForCronJob()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("*/2 * * * *").WithName("JobName")));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var successful = registry.TryGetSchedule("JobName", out var cronExpression, out var timeZoneInfo);
+
+        successful.ShouldBeTrue();
+        cronExpression.ShouldBe("*/2 * * * *");
+        timeZoneInfo.ShouldBe(TimeZoneInfo.Utc);
+    }
+
+    [Fact]
+    public async Task ShouldReturnFalseIfGivenJobWasNotFound()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob();
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var successful = registry.TryGetSchedule("JobName", out var cronExpression, out var timeZoneInfo);
+
+        successful.ShouldBeFalse();
+        cronExpression.ShouldBeNull();
+        timeZoneInfo.ShouldBeNull();
+    }
+
+    [Fact]
+    public async Task ShouldFindDependentJobsWithAGivenName()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s =>
+        {
+            s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("Job1"))
+                .ExecuteWhen(r => r.RunJob(() => { }, "Job2"));
+        });
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var successful = registry.TryGetSchedule("Job2", out var cronExpression, out var timeZoneInfo);
+        successful.ShouldBeTrue();
+        cronExpression.ShouldBeNull();
+        timeZoneInfo.ShouldBeNull();
+    }
+
+    [Fact]
+    public async Task UpdatingParameterHasImmediateEffect()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p
+            .WithCronExpression("* * * * *")
+            .WithParameter("foo")
+            .WithName("JobName")));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.UpdateParameter("JobName", "Bar");
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var content = await CommunicationChannel.Reader.ReadAsync(CancellationToken);
+        content.ShouldBe("Bar");
+    }
+
+    [Fact]
+    public void ShouldRetrieveAllSchedules()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        var timeZone = TimeZoneInfo.FindSystemTimeZoneById("Pacific Standard Time");
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p
+            .WithCronExpression("*/2 * * * *", timeZoneInfo: timeZone)
+            .WithName("JobName")));
+        var provider = CreateServiceProvider();
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+        registry.AddJob(s => s.AddJob(() => { }, "* * * * *", jobName: "JobName2"));
+
+        var allSchedules = registry.GetAllRecurringJobs();
+
+        allSchedules.Count.ShouldBe(2);
+        allSchedules.ShouldContain(s => s.JobName == "JobName"
+                                        && s.CronExpression == "*/2 * * * *"
+                                        && s.JobType == typeof(SimpleJob)
+                                        && s.TimeZone == timeZone);
+        allSchedules.ShouldContain(s => s.JobName == "JobName2"
+                                        && s.CronExpression == "* * * * *"
+                                        && s.TimeZone == TimeZoneInfo.Utc
+                                        && s.JobType == null);
+    }
+
+    [Fact]
+    public void AddingJobDuringRuntimeIsRetrieved()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(p => p.AddJob<SimpleJob>());
+        var provider = CreateServiceProvider();
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+        registry.AddJob(n => n.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+
+        var allSchedules = registry.GetAllRecurringJobs();
+
+        allSchedules.Count.ShouldBe(1);
+        allSchedules.ShouldContain(s => s.JobName == "JobName"
+                                        && s.CronExpression == "* * * * *"
+                                        && s.TimeZone == TimeZoneInfo.Utc
+                                        && s.JobType == typeof(SimpleJob));
+    }
+
+    [Fact]
+    public async Task ShouldDisableJob()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        registry.DisableJob("JobName");
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(1, TimeSpan.FromMilliseconds(200));
+        jobFinished.ShouldBeFalse();
+    }
+
+    [Fact]
+    public async Task ShouldThrowAnExceptionWhenJobIsNotFoundAndTryingToDisable()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob();
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        Should.Throw<InvalidOperationException>(() => registry.DisableJob("JobName"));
+    }
+
+    [Fact]
+    public async Task ShouldEnableJob()
+    {
+        var fakeTimer = new FakeTimeProvider();
+        ServiceCollection.AddSingleton<TimeProvider>(fakeTimer);
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+        registry.DisableJob("JobName");
+
+        registry.EnableJob("JobName");
+
+        fakeTimer.Advance(TimeSpan.FromMinutes(1));
+        var jobFinished = await WaitForJobsOrTimeout(1);
+        jobFinished.ShouldBeTrue();
+    }
+
+    [Fact]
+    public async Task ShouldThrowWhenDuplicateJobNamesDuringRegistration()
+    {
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")
+            .And.WithCronExpression("*/2 * * * *").WithName("JobName")));
+
+        var act = async () => await CreateServiceProvider().GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+
+        await act.ShouldThrowAsync<InvalidOperationException>();
+    }
+
+    [Fact]
+    public void ShouldThrowRuntimeExceptionWithDuplicateJob()
+    {
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        var runtimeJobRegistry = CreateServiceProvider().GetRequiredService<IRuntimeJobRegistry>();
+
+        var act = () => runtimeJobRegistry.AddJob(s => s.AddJob(() =>
+        {
+        }, "* * * * *", jobName: "JobName"));
+
+        act.ShouldThrow<InvalidOperationException>();
+    }
+
+    private sealed class SimpleJob(ChannelWriter<object> writer) : IJob
+    {
+        public async Task RunAsync(JobExecutionContext context, CancellationToken token) => await writer.WriteAsync(context.Parameter ?? string.Empty, token);
+    }
+}